single-file module
