--- crates/engine/tree/src/tree/root/mod.rs
@@ -1,45 +1,0 @@
-//! State root computation related code.
-
-mod task;
-
-#[cfg(test)]
-mod tests;
-
-use reth_provider::providers::ConsistentDbView;
-use reth_trie::{updates::TrieUpdates, TrieInput};
-use reth_trie_parallel::root::ParallelStateRootError;
-use revm_primitives::B256;
-use std::sync::{mpsc, Arc};
-
-/// Result of the state root calculation
-pub(crate) type StateRootResult = Result<(B256, TrieUpdates), ParallelStateRootError>;
-
-/// Handle to a spawned state root task.
-#[derive(Debug)]
-#[allow(dead_code)]
-pub(crate) struct StateRootHandle {
-    /// Channel for receiving the final result.
-    rx: mpsc::Receiver<StateRootResult>,
-}
-
-#[allow(dead_code)]
-impl StateRootHandle {
-    /// Creates a new handle from a receiver.
-    pub(crate) const fn new(rx: mpsc::Receiver<StateRootResult>) -> Self {
-        Self { rx }
-    }
-
-    /// Waits for the state root calculation to complete.
-    pub(crate) fn wait_for_result(self) -> StateRootResult {
-        self.rx.recv().expect("state root task was dropped without sending result")
-    }
-}
-
-/// Common configuration for state root tasks
-#[derive(Debug)]
-pub(crate) struct StateRootConfig<Factory> {
-    /// View over the state in the database.
-    pub consistent_view: ConsistentDbView<Factory>,
-    /// Latest trie input.
-    pub input: Arc<TrieInput>,
-}
